Enable tridactyl-native in FF
After many hours of debugging and soul searching, it turns out that tridactyl-native simply didn't work because unveil did not allow the file to be read or executed. OpenBSD is too secure I swear

While native now works, text-in-editor and context menus still do not work because tridactyl does not have access to the display :0. Not sure if it's a bug or another security safe measure.